+++ plugin-dev/Source/Sentry/Private/SentrySettings.cpp
@@ -12,12 +12,13 @@ USentrySettings::USentrySettings(const FObjectInitializer& ObjectInitializer)
 	, EnableVerboseLogging(true)
 	, EnableAutoCrashCapturing(true)
 	, EnableAutoLogAttachment(false)
+	, EnableAutoSessionTracking(true)
+	, SessionTimeout(30000)
+	, OverrideReleaseName(false)
 	, UploadSymbolsAutomatically(false)
 	, IncludeSources(false)
 	, CrashReporterUrl()
 {
-	GConfig->GetString(TEXT("/Script/EngineSettings.GeneralProjectSettings"), TEXT("ProjectVersion"), Release, GGameIni);
-
 #if WITH_EDITOR
 	Environment = TEXT("Editor");
 #elif UE_BUILD_SHIPPING