+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Add release health monitoring feature ([#270](https://github.com/getsentry/sentry-unreal/pull/270))
+
 ### Dependencies
 
 - Bump Native SDK from v0.6.1 to v0.6.2 ([#269](https://github.com/getsentry/sentry-unreal/pull/269))